Beta update
Working bases with security, please edit the "piwebctrl.py" for the password.

+++ piwebctrl.py
@@ -5,13 +5,20 @@ from http.server import BaseHTTPRequestHandler, HTTPServer
 import time
 import os
 from os import path
+import urllib.parse as urlparse
+from urllib.parse import parse_qs
 
 hostName = ""
 hostPort = 80
+password = "42758"
 
 class piWebCtrl(BaseHTTPRequestHandler):
 
     #GET is for clients geting the predi
+    def getPass(self,url):
+        parsed = urlparse.urlparse("http://localhost"+url)
+        return str(parse_qs(parsed.query)['pass']).replace("['","").replace("']","")
+        
     def do_GET(self):
         thisPath = self.path.split("?")
         if thisPath[0] == "/":
@@ -20,13 +27,21 @@ class piWebCtrl(BaseHTTPRequestHandler):
             fstr = open(r"web/index.html", encoding="utf-8").read()
             self.wfile.write(bytes(fstr, "utf-8"))
         elif thisPath[0] == "/run/reboot":
-            self.send_response(200)
-            self.wfile.write(bytes('{"html":"Rebooting the Raspberry Pi","cmd":null}', "utf-8"))
-            os.system("sudo reboot &")
+            if self.getPass(self.path) == password:
+                self.send_response(200)
+                self.wfile.write(bytes('{"html":"Rebooting the Raspberry Pi","cmd":null}', "utf-8"))
+                os.system("sudo reboot &")
+            else:
+                self.send_response(200)
+                self.wfile.write(bytes('{"html":"Bad password, please retry.","cmd":null}', "utf-8"))
         elif thisPath[0] == "/run/poweroff":
-            self.send_response(200)
-            self.wfile.write(bytes('{"html":"Powering off the Raspberry Pi","cmd":null}', "utf-8"))
-            os.system("sudo poweroff &")
+            if self.getPass(self.path) == password:
+                self.send_response(200)
+                self.wfile.write(bytes('{"html":"Powering off the Raspberry Pi","cmd":null}', "utf-8"))
+                os.system("sudo poweroff &")
+            else:
+                self.send_response(200)
+                self.wfile.write(bytes('{"html":"Bad password, please retry.","cmd":null}', "utf-8"))
         else:
             if path.exists("web"+thisPath[0]) is True:
                 self.send_response(200)
